+++ backend/src/main/java/com/festago/bookmark/application/FestivalBookmarkV1QueryService.java
@@ -2,7 +2,7 @@ package com.festago.bookmark.application;
 
 import com.festago.bookmark.dto.v1.FestivalBookmarkV1Response;
 import com.festago.bookmark.repository.FestivalBookmarkOrder;
-import com.festago.bookmark.repository.FestivalBookmarkV1QueryDslRepository;
+import com.festago.bookmark.repository.v1.FestivalBookmarkV1QueryDslRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
